Migrate entire project from JavaScript to TypeScript
- Convert all source files to TypeScript with strict mode (no implicit any)
- Migrate from CommonJS to ES Modules with NodeNext resolution
- Create src/ directory structure with types/ subdirectory
- Migrate test files to TypeScript
- Upgrade dependencies:
  - sqlite: 3.0.0 -> ^5.1.1 (proper types)
  - make-dir: ^5.1.0 (ESM support)
  - untildify: ^6.0.0 (ESM support)
  - Add sqlite3 ^5.1.7 as peer dependency
- Add tsconfig.json with strict settings
- Update package.json with proper ESM exports and types
- Add dist/ to .gitignore for compiled output
- Update CLAUDE.md with new architecture documentation
