+++ package.json
@@ -1,12 +1,28 @@
 {
   "name": "backup-bear-notes",
-  "version": "1.0.4",
+  "version": "1.0.5",
   "description": "Back up your Bear Notes content to a folder",
-  "bin": "index.js",
+  "type": "module",
+  "bin": "./dist/index.js",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    }
+  },
+  "files": [
+    "dist"
+  ],
   "scripts": {
+    "backup": "tsx src/index.ts",
+    "build": "tsc",
+    "prepublishOnly": "npm run build",
     "test": "vitest run",
     "test:watch": "vitest",
-    "test:coverage": "vitest run --coverage"
+    "test:coverage": "vitest run --coverage",
+    "typecheck": "tsc --noEmit"
   },
   "repository": {
     "type": "git",
@@ -21,14 +37,18 @@
   "author": "TehShrike",
   "license": "WTFPL",
   "dependencies": {
-    "make-dir": "1.3.0",
-    "mri": "^1.1.4",
-    "pify": "4.0.1",
-    "sqlite": "3.0.0",
-    "untildify": "3.0.3"
+    "make-dir": "^5.1.0",
+    "mri": "^1.2.0",
+    "sqlite": "^5.1.1",
+    "sqlite3": "^5.1.7",
+    "untildify": "^6.0.0"
   },
   "devDependencies": {
+    "@types/node": "^22.0.0",
+    "@types/sqlite3": "^3.1.11",
     "@vitest/coverage-v8": "^2.1.0",
+    "tsx": "^4.7.0",
+    "typescript": "^5.7.0",
     "vitest": "^2.1.0"
   }
 }
